refactor: 설문 관리 DTO inner class로 변경
