+++ survwey/survwey/src/main/java/mcnc/survwey/api/survey/manage/dto/QuestionDTO.java
@@ -1,12 +1,16 @@
 package mcnc.survwey.api.survey.manage.dto;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import mcnc.survwey.domain.question.enums.QuestionType;
 import mcnc.survwey.domain.question.Question;
+import mcnc.survwey.domain.selection.Selection;
+import mcnc.survwey.domain.selection.SelectionId;
 import mcnc.survwey.domain.survey.Survey;
 
 
@@ -21,21 +25,21 @@ public class QuestionDTO {
     private Long quesId;
     @NotBlank(message = "질문 내용은 필수입니다.")
     private String body;
-
     private QuestionType questionType;
-
     private List<SelectionDTO> selectionList;
 
     public static QuestionDTO of(Question question) {
-        List<SelectionDTO> selectionDTOList = question.getSelectionList()
-                .stream().map(SelectionDTO::of).toList();
-
         return QuestionDTO.builder()
                 .quesId(question.getQuesId())
                 .body(question.getBody())
                 .questionType(question.getType())
-                .selectionList(selectionDTOList)
+                .selectionList(SelectionDTO.ofList(question.getSelectionList()))
                 .build();
+    }
+
+    public static List<QuestionDTO> ofList(List<Question> questionList) {
+        return questionList.stream()
+                .map(QuestionDTO::of).toList();
     }
 
     public Question toEntity(Survey survey) {
@@ -45,4 +49,44 @@ public class QuestionDTO {
                 .survey(survey)
                 .build();
     }
+
+
+    @Data
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @Builder
+    public static class SelectionDTO {
+
+        private SelectionId selectionId;
+
+        @NotBlank(message = "보기 내용은 필수입니다.")
+        private String body;
+
+        @JsonProperty("isEtc")
+        private boolean isEtc;
+
+        public static SelectionDTO of(Selection selection) {
+            return SelectionDTO.builder()
+                    .selectionId(selection.getId())
+                    .isEtc(selection.isEtc())
+                    .body(selection.getBody())
+                    .build();
+        }
+
+        public static List<SelectionDTO> ofList(List<Selection> selectionList) {
+            return selectionList.stream()
+                    .map(SelectionDTO::of)
+                    .toList();
+        }
+
+        public Selection toEntity(SelectionId selectionId, Question question) {
+            return Selection.builder()
+                    .id(selectionId)
+                    .body(this.getBody())
+                    .isEtc(this.isEtc())
+                    .question(question)
+                    .build();
+        }
+    }
+
 }
